requirements.txt
sqlalchemy>=2.0.0
typer>=0.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi[standard]>=0.115.0
jinja2>=3.1.0
sse-starlette>=2.0.0
